requirements.txt
Django==5.0.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
pillow==10.2.0
